quickstart/requirements.txt
python-dotenv
pipecat-ai[daily,google,silero]
